indigo_token changed to long long unsigned

+++ indigo_libs/indigo/indigo_token.h
@@ -36,7 +36,7 @@
 extern "C" {
 #endif
 
-typedef uint64_t indigo_token;
+typedef long long unsigned indigo_token;
 
 
 /** Convert hex string to indigo_token, 0 for error or no token
